# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DTH Customer Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
}

export interface Provider {
  id: string;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  defaultValidityDays: number;
}

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  active: 'active',
  expiring_soon: 'expiring_soon',
  expired: 'expired',
  pending: 'pending',
} as const;

export interface Customer {
  id: number;
  name: string;
  mobile: string;
  /** @nullable */
  address?: string | null;
  provider: string;
  customerId: string;
  status: CustomerStatus;
  /** @nullable */
  lastRechargeDate?: string | null;
  /** @nullable */
  nextRechargeDate?: string | null;
  /** @nullable */
  lastRechargeAmountInr?: number | null;
  /** @nullable */
  lastRechargeAmountLkr?: number | null;
  /** @nullable */
  daysUntilExpiry?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CustomerDetailStatus = typeof CustomerDetailStatus[keyof typeof CustomerDetailStatus];


export const CustomerDetailStatus = {
  active: 'active',
  expiring_soon: 'expiring_soon',
  expired: 'expired',
  pending: 'pending',
} as const;

export interface Recharge {
  id: number;
  customerId: number;
  rechargeDate: string;
  validityDays: number;
  nextRechargeDate: string;
  amountInr: number;
  amountLkr: number;
  /** @nullable */
  planName?: string | null;
  /** @nullable */
  profitMargin?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CustomerDetail {
  id: number;
  name: string;
  mobile: string;
  /** @nullable */
  address?: string | null;
  provider: string;
  customerId: string;
  status: CustomerDetailStatus;
  /** @nullable */
  lastRechargeDate?: string | null;
  /** @nullable */
  nextRechargeDate?: string | null;
  /** @nullable */
  lastRechargeAmountInr?: number | null;
  /** @nullable */
  lastRechargeAmountLkr?: number | null;
  /** @nullable */
  daysUntilExpiry?: number | null;
  createdAt: string;
  updatedAt: string;
  recharges: Recharge[];
}

export interface CustomerInput {
  name: string;
  mobile: string;
  address?: string;
  provider: string;
  customerId: string;
}

export interface CustomerUpdate {
  name?: string;
  mobile?: string;
  address?: string;
  provider?: string;
  customerId?: string;
}

export interface RechargeInput {
  /** YYYY-MM-DD format */
  rechargeDate: string;
  /** Number of days the recharge is valid */
  validityDays: number;
  /** Amount in Indian Rupees */
  amountInr: number;
  /** Equivalent amount in Sri Lankan Rupees */
  amountLkr: number;
  planName?: string;
  profitMargin?: number;
  notes?: string;
  sendSms?: boolean;
}

export interface ProviderCount {
  provider: string;
  count: number;
  activeCount: number;
}

export interface DashboardStats {
  totalCustomers: number;
  activeCustomers: number;
  expiringSoonCustomers: number;
  expiredCustomers: number;
  pendingCustomers: number;
  monthlyRevenueInr: number;
  monthlyRevenueLkr: number;
  totalRevenueInr: number;
  totalRevenueLkr: number;
  customersByProvider: ProviderCount[];
}

export interface DueAlerts {
  dueToday: Customer[];
  dueIn3Days: Customer[];
  dueIn7Days: Customer[];
  dueIn15Days: Customer[];
  overdue: Customer[];
}

export interface SmsInput {
  customerId: number;
  message: string;
  mobile: string;
}

export type ListCustomersParams = {
provider?: string;
status?: ListCustomersStatus;
search?: string;
/**
 * Filter by customers due within N days
 */
dueDays?: number;
};

export type ListCustomersStatus = typeof ListCustomersStatus[keyof typeof ListCustomersStatus];


export const ListCustomersStatus = {
  active: 'active',
  expiring_soon: 'expiring_soon',
  expired: 'expired',
  pending: 'pending',
} as const;

export type GetDueAlertsParams = {
days?: number;
};