# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DTH Customer Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: customers
    description: Customer management
  - name: recharges
    description: Recharge management
  - name: dashboard
    description: Dashboard stats
  - name: providers
    description: DTH providers

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all DTH providers
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers with optional filters
      parameters:
        - name: provider
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expiring_soon, expired, pending]
        - name: search
          in: query
          schema:
            type: string
        - name: dueDays
          in: query
          description: Filter by customers due within N days
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /customers/{customerId}/recharges:
    get:
      operationId: listRecharges
      tags: [recharges]
      summary: Get recharge history for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of recharges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recharge"
    post:
      operationId: addRecharge
      tags: [recharges]
      summary: Add a recharge for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "201":
          description: Recharge added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recharge"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recharges/{id}:
    delete:
      operationId: deleteRecharge
      tags: [recharges]
      summary: Delete a recharge record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Recharge deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard overview statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/due-alerts:
    get:
      operationId: getDueAlerts
      tags: [dashboard]
      summary: Get customers due within specified days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Due customers grouped by urgency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DueAlerts"

  /customers/refresh-status:
    post:
      operationId: refreshAllStatuses
      tags: [customers]
      summary: Recalculate and refresh statuses for all customers
      responses:
        "200":
          description: Statuses refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /sms/send:
    post:
      operationId: sendSms
      tags: [customers]
      summary: Send SMS notification to a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsInput"
      responses:
        "200":
          description: SMS sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: ["string", "null"]
      required:
        - authenticated
        - username

    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        defaultValidityDays:
          type: integer
      required:
        - id
        - name
        - defaultValidityDays

    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        address:
          type: ["string", "null"]
        provider:
          type: string
        customerId:
          type: string
        status:
          type: string
          enum: [active, expiring_soon, expired, pending]
        lastRechargeDate:
          type: ["string", "null"]
        nextRechargeDate:
          type: ["string", "null"]
        lastRechargeAmountInr:
          type: ["number", "null"]
        lastRechargeAmountLkr:
          type: ["number", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - mobile
        - provider
        - customerId
        - status
        - createdAt
        - updatedAt

    CustomerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        address:
          type: ["string", "null"]
        provider:
          type: string
        customerId:
          type: string
        status:
          type: string
          enum: [active, expiring_soon, expired, pending]
        lastRechargeDate:
          type: ["string", "null"]
        nextRechargeDate:
          type: ["string", "null"]
        lastRechargeAmountInr:
          type: ["number", "null"]
        lastRechargeAmountLkr:
          type: ["number", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        recharges:
          type: array
          items:
            $ref: "#/components/schemas/Recharge"
      required:
        - id
        - name
        - mobile
        - provider
        - customerId
        - status
        - createdAt
        - updatedAt
        - recharges

    CustomerInput:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        address:
          type: string
        provider:
          type: string
        customerId:
          type: string
      required:
        - name
        - mobile
        - provider
        - customerId

    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        address:
          type: string
        provider:
          type: string
        customerId:
          type: string

    Recharge:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        rechargeDate:
          type: string
        validityDays:
          type: integer
        nextRechargeDate:
          type: string
        amountInr:
          type: number
        amountLkr:
          type: number
        planName:
          type: ["string", "null"]
        profitMargin:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerId
        - rechargeDate
        - validityDays
        - nextRechargeDate
        - amountInr
        - amountLkr
        - createdAt

    RechargeInput:
      type: object
      properties:
        rechargeDate:
          type: string
          description: YYYY-MM-DD format
        validityDays:
          type: integer
          description: Number of days the recharge is valid
        amountInr:
          type: number
          description: Amount in Indian Rupees
        amountLkr:
          type: number
          description: Equivalent amount in Sri Lankan Rupees
        planName:
          type: string
        profitMargin:
          type: number
        notes:
          type: string
        sendSms:
          type: boolean
          default: false
      required:
        - rechargeDate
        - validityDays
        - amountInr
        - amountLkr

    DashboardStats:
      type: object
      properties:
        totalCustomers:
          type: integer
        activeCustomers:
          type: integer
        expiringSoonCustomers:
          type: integer
        expiredCustomers:
          type: integer
        pendingCustomers:
          type: integer
        monthlyRevenueInr:
          type: number
        monthlyRevenueLkr:
          type: number
        totalRevenueInr:
          type: number
        totalRevenueLkr:
          type: number
        customersByProvider:
          type: array
          items:
            $ref: "#/components/schemas/ProviderCount"
      required:
        - totalCustomers
        - activeCustomers
        - expiringSoonCustomers
        - expiredCustomers
        - pendingCustomers
        - monthlyRevenueInr
        - monthlyRevenueLkr
        - totalRevenueInr
        - totalRevenueLkr
        - customersByProvider

    ProviderCount:
      type: object
      properties:
        provider:
          type: string
        count:
          type: integer
        activeCount:
          type: integer
      required:
        - provider
        - count
        - activeCount

    DueAlerts:
      type: object
      properties:
        dueToday:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        dueIn3Days:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        dueIn7Days:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        dueIn15Days:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        overdue:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
      required:
        - dueToday
        - dueIn3Days
        - dueIn7Days
        - dueIn15Days
        - overdue

    SmsInput:
      type: object
      properties:
        customerId:
          type: integer
        message:
          type: string
        mobile:
          type: string
      required:
        - customerId
        - message
        - mobile